Extract FrameEffect typeclass
